+++ spack/packages/dla-future-fortran/package.py
@@ -46,7 +46,11 @@ class DlaFutureFortran(CMakePackage):
         args = []
 
         args.append(self.define_from_variant("BUILD_SHARED_LIBS", "shared"))
-        args.append(self.define_from_variant("DLAF_FORTRAN_BUILD_TESTING", "test"))
+
+        if self.spec.satisfies("+test"):
+            args.append(self.define("DLAF_FORTRAN_BUILD_TESTING", True))
+            # Tests run with 6 MPI ranks
+            args.append(self.define("MPIEXEC_MAX_NUMPROCS", 6))
 
         # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
         # FIXME: Variables only available on the DLA-Future-Fortran repo