+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## DLA-Future-Fortran X.Y.Z
+
+### Fixed
+
+* Spack installation with `+test` variant by setting `-DMPIEXEC_MAX_NUMPROCS=6` [PR #10]
+
 ## DLA-Future-Fortran 0.1.0
 
 First release of [DLA-Future-Fortran], a Fortran interface for [DLA-Future].